+++ src/pages/login/types/auth.ts
@@ -10,11 +10,6 @@ export interface KakaoLoginResponse {
   prefill: KakaoLoginPrefill | null;
 }
 
-// 로그아웃 관련 타입
-export interface LogoutResponse {
-  message: string;
-}
-
 // 로그인 API 응답 타입
 export interface LoginApiResponse {
   data: KakaoLoginResponse;
