+++ src/pages/mypage/types/apis/auth.ts
@@ -1,0 +1,4 @@
+// 로그아웃 관련 타입
+export interface LogoutResponse {
+  message: string;
+}
